Extract button into separate component
