+++ frontend/src/app/pages/exercises/exercise/shared/start-pause-button/start-pause-button.component.html
@@ -1,0 +1,16 @@
+<button
+    *ngIf="(exerciseStatus$ | async) === 'running'"
+    (click)="pauseExercise()"
+    class="btn btn-primary ms-3"
+    data-cy="trainerToolbarPauseButton"
+>
+    Pause
+</button>
+<button
+    *ngIf="(exerciseStatus$ | async) !== 'running'"
+    (click)="startExercise()"
+    class="btn btn-primary ms-3"
+    data-cy="trainerToolbarStartButton"
+>
+    Start
+</button>
