+++ frontend/src/app/pages/exercises/exercise/shared/trainer-toolbar/trainer-toolbar.component.html
@@ -110,22 +110,7 @@
         </div>
     </div>
 
-    <button
-        *ngIf="(exerciseStatus$ | async) === 'running'"
-        (click)="pauseExercise()"
-        class="btn btn-primary ms-3"
-        data-cy="trainerToolbarPauseButton"
-    >
-        Pause
-    </button>
-    <button
-        *ngIf="(exerciseStatus$ | async) !== 'running'"
-        (click)="startExercise()"
-        class="btn btn-primary ms-3"
-        data-cy="trainerToolbarStartButton"
-    >
-        Start
-    </button>
+    <app-start-pause-button />
     <button (click)="deleteExercise()" class="btn btn-danger ms-3 float-end">
         Übung löschen
     </button>
